Move byte pattern finder to utils
